scripts/youtube-knowledge-base/requirements.txt
browser-cookie3>=0.19.0
spacy>=3.8.0
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.32.0
